+++ ui/src/app/core/job-manager.service.ts
@@ -8,7 +8,7 @@ import {QueryJobsResponse} from '../shared/model/QueryJobsResponse';
 import {JobMetadataResponse} from '../shared/model/JobMetadataResponse';
 import {UpdateJobLabelsRequest} from "../shared/model/UpdateJobLabelsRequest";
 import {UpdateJobLabelsResponse} from "../shared/model/UpdateJobLabelsResponse";
-import {TimeFrame, AggregationResponse} from '../shared/model/models';
+import {TimeFrame, AggregationResponse, JobAttemptsResponse} from '../shared/model/models';
 
 import {ConfigLoaderService} from "../../environments/config-loader.service";
 
@@ -159,5 +159,25 @@ export class JobManagerService {
       .toPromise()
       .then(response => this.convertToAggregationJobsResponse(response.json()))
       .catch((e) => this.handleError(e));
+  }
+
+  getTaskAttempts(id:string, task:string): Promise<JobAttemptsResponse> {
+    const apiUrl = this.configLoader.getEnvironmentConfigSynchronous()['apiUrl'];
+    return this.http.get(`${apiUrl}/jobs/${id}/${task}/attempts`,
+      new RequestOptions({headers: this.getHttpHeaders()}))
+      .toPromise()
+      .then(response => response.json())
+      .catch((e) => this.handleError(e));
+
+  }
+  //
+  getShardAttempts(id:string, task:string, index:number): Promise<JobAttemptsResponse> {
+    const apiUrl = this.configLoader.getEnvironmentConfigSynchronous()['apiUrl'];
+    return this.http.get(`${apiUrl}/jobs/${id}/${task}/${index}/attempts`,
+      new RequestOptions({headers: this.getHttpHeaders()}))
+      .toPromise()
+      .then(response => response.json())
+      .catch((e) => this.handleError(e));
+
   }
 }