+++ ui/src/app/job-details/job-details.component.spec.ts
@@ -6,6 +6,7 @@ import {Component, DebugElement} from '@angular/core';
 import {
   MatButtonModule,
   MatCardModule,
+  MatDialogModule,
   MatDividerModule,
   MatExpansionModule,
   MatListModule,
@@ -34,6 +35,8 @@ import {JobStatus} from "../shared/model/JobStatus";
 import {JobMetadataResponse} from '../shared/model/JobMetadataResponse';
 import {JobFailuresTableComponent} from "./common/failures-table/failures-table.component";
 import {JobTimingDiagramComponent} from "./tabs/timing-diagram/timing-diagram.component";
+import {JobScatteredAttemptsComponent} from "./tabs/scattered-attempts/scattered-attempts.component";
+import {JobAttemptComponent} from "./common/attempt/attempt.component";
 
 describe('JobDetailsComponent', () => {
 
@@ -60,6 +63,7 @@ describe('JobDetailsComponent', () => {
         AppComponent,
         FakeJobListComponent,
         TestJobDetailsComponent,
+        JobAttemptComponent,
         JobDebugIconsComponent,
         JobDetailsComponent,
         JobFailuresTableComponent,
@@ -68,6 +72,7 @@ describe('JobDetailsComponent', () => {
         JobResourcesTableComponent,
         JobTabsComponent,
         JobTimingDiagramComponent,
+        JobScatteredAttemptsComponent,
       ],
       imports: [
         ClrIconModule,
@@ -75,6 +80,7 @@ describe('JobDetailsComponent', () => {
         CommonModule,
         MatButtonModule,
         MatCardModule,
+        MatDialogModule,
         MatDividerModule,
         MatExpansionModule,
         MatListModule,