+++ ui/src/app/shared/common.ts
@@ -62,3 +62,8 @@ export const defaultTimeFrame = TimeFrame.DAYS7;
 /** The page size to initially request from the backend. The maximum number of
  *  jobs on a page is 100, so request 200 initially. */
 export const initialBackendPageSize = 200;
+
+
+export function objectNotEmpty(object: object): boolean {
+  return object && Object.keys(object).length !== 0;
+}