+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Job Manager Change Log
 
+## v1.0.0 Release Notes
+
+### Multiple changes to the "List" tab in the Job Details page for Cromwell jobs.
+
+#### If a task or shard has multiple attempts it can be clicked on, which will expand to show additional rows, one for each attempt.
+
+#### Scattered task names are a link that will display scattered task shards in a dialog box.
+
+#### Scattered sub-workflow shards are a link that will take the users to a "Job Details" page with details of that shard.
+
+#### Added icons for inputs and outputs which can be clicked on to show the full list.
+
+#### Changed the call-cached icon and moved it into the "duration" column.
+
+### Re-implemented shard data requests to get data more efficiently from the API.
+
 ## v0.7.3 Release Notes
 
 ### Fixed the broken display of icons in Firefox.